move indicator_upgrade logic
