+++ etl/indicator_upgrade/indicator_update.py
@@ -12,13 +12,13 @@ from sqlalchemy.orm import Session
 from structlog import get_logger
 
 import etl.grapher_model as gm
-from etl.chart_revision.v3.schema import (
+from etl.db import get_engine
+from etl.indicator_upgrade.schema import (
     fix_errors_in_schema,
     get_schema_chart_config,
     validate_chart_config_and_remove_defaults,
     validate_chart_config_and_set_defaults,
 )
-from etl.db import get_engine
 
 # Logger
 log = get_logger()
